requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pgvector==0.2.5
jinja2==3.1.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
websockets==12.0
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
numpy==1.26.4
PyJWT==2.8.0
bcrypt==3.2.2